requirements.txt
dataclasses
openpyxl
python-dotenv
requests
simple_salesforce